+++ internal/adapters/settings.go
@@ -2,5 +2,4 @@ package adapters
 
 const (
 	CServicePrefix = "hidden-lake-adapter"
-	CVoidHost      = "_"
 )